refactor - separate into class
